Remove cache-control header override
We were overriding the cache-control header to set a 5m private cache to any
API responses deemed 'cachable' (if they serve an ETag). This was done because
it was not possible/straight forward at the time to set cache-control response
headers in the API due to the requests being authenticated (and ASP.NET firmly
preventing us).

This is no longer the case and we are implementing the cache-control headers in
the API, so they can be removed from the API client.