Explicitly format Date/Time attributes in requests
Currently any `DateTime` or `Time` attributes are serialised to JSON by calling
`to_s` which, in turn, calls `to_formatted_s(:default)`. The issue with this is
that the default formatting behaves differently for `DateTime` or `Time`
objects that have a `UTC` offset (as opposed to a `+00:00` offset). It results
in `UTC` being present in the formatted date in place of the timezone offset,
which isn't supported by the API.

This commit intercepts the serialisation process for both the JSON request body
and the query string parameters, explicitly formatting them to ensure the
offset is output in the `+/-HH:MM` format.