Format Date correctly in request body/query string
After updating the gem to include the latest API methods for some reason the
`Date` types started formatting differently when calling `strftime` with a date
(and time) format -- they changed from including only the date portion:

```
2000-01-24
```

To having the time added onto the end:

```
2000-01-24 00:00:00 +00:00
```

The API rejects this as a valid `Date` value. This commit updates the API
client monkey patch to format `Date` objects explicitly.