Generate with latest swagger-codegen
The version of `swagger-codegen` we have been using until now generated the
code with `URI.encode`, which has been deprecated in Ruby 2.7 -- a newer
version of `swagger-codegen` replaces this with `Addressable::URI.encode`.

A side-effect of this upgrade is that the specs fail due to trying to encode a
url with no host; `URI.encode` didn't complain, but `Addressable::URI.encode`
raises an error, so we need to manually udpate the specs to set an example host
before the tests will pass.

It would be better if we didn't have to do this step, but I can't see a way
around it short of wrapping the auto-generated gem specs in our own rake task
that monkey patches `GetIntoTeachingApiClient::Configuration` with a default
host... but then the gem's specs won't run independently and I think that would
be confusing for people coming to the project.