Support request_id in Thread.current
We want to be able to track requests from the app right through to the
API. In order to do this we need to set a correlation id in the request
headers going to the API. Rails has a request id we can tap into, but
getting it into the API client for all requests will be
difficult/verbose.

To make life easier this commit updates the API client to expect a
request id set on the current thread. We are using
`ActiveSupport::CurrentAttributes` to ensure the thread globals get
cleaned up after each request.

If a `request_id` is present it will be written to the `Request-Id`
header prior to making the API request (ASP.NET Core will pick up on
this and use it as the `TraceId`).