Regenerate with latest swagger-codegen
We were previously using `swagger-codegen@2` to generate the API client code
from the Swagger specification; in the past the recent versions of
`swagger-codegen` did not support Ruby. As this has since changed, we can
update to the latest version and regenerate the client.

Fixes an issue where v2 would generate initializers that accepted `camelCase`
attributes instead of `snake_case`.

Update README with fixes for issues in the latest `swagger-codegen` output;
namely the `base_object_spec.rb` has hard coded references to an example schema
and `problem_details.rb` has a call to `super(attributes)` when it doesn't have
a specific super class that adheres to that interface.

Bumps the version to 2.0.0.