Regenerate using V3 of the Swagger OpenApi
We previously used V2 but its not generating some of the types accurately in
the latest swagger-codegen version (they are being mapped to the wrong type;
string instead of integer in places).

To try and get around this I've configured the API to generate v3 swagger docs
and re-ran against that output.