Regnerate using V3 OpenApi format and openapi-generator
Previously we were using `swagger-codegen@2` for generating the API client;
this mostly worked fine, but requried patching to use Faraday (for HTTP
caching) and various other monkey patches/tweaks to customise functionality.
The latest version of `swagger-codegen` fixes a bug where initializers were
`camelCase` instead of `snake_case`, however it had other bugs (such as every
type was classed as an `Object`!).

This commit regenerates the API client using `openapi-generator`, which is
compatible with v3 of the Open API/Swagger format and appears to do a decent
job of generating the client in Ruby with these advantages:

- Correct initializers in `snake_case`
- Correct typing throughout
- Uses Faraday without monkey patching
- Additional validation rules appear to get mapped into the models
- Models raise if initialized with unknown attributes

I've left two previous commits in the resulting PR to show the output from
other attempts.