Add default request timeout of 10 seconds
We are seeing timeout errors in the School Experience app; I believe this is
because the GSE timeout is 30 seconds, which is shorter than the API client
timeout (defaults to 60 seconds). This results in the Faraday retry logic never
being able to execute, as the originating request times out before it has
chance.

Set a default request timeout for the API client of 10 seconds and increase the
number of retries to 2. This way at least one timeout/retry will happen before
the downstream clients timeout the original request.