Update faraday to 2.x
Apps using this gem and have their own faraday or use other gems
that use faraday, were struggling with version conflicts.

This gem uses an old version of faraday, 1.x. Whilst other libraries or
apps use 2.x.

So we need to upgrade faraday, and in doing so we removed
faraday-middleware as it's deprecated and according to the readme of the
repository, a lot of the functionality has been incorporated by the
faraday gem itself, in 2.x