Changed the CacheInvalidator spec
With the new new updates, of faraday, ruby and the generator.
The cache cannot be cleared so easily in specs anymore.

This doesn't mean the configuration doesn't work. When we set a
cache_store it's still the cache_store faraday will use but the instance
in the spec is not the most up to date one. So calling cache_store.clear is not working.

So each spec that tests the CacheInvalidator needs to hit a different endpoint.
This way specs don't write to the cache for other specs.

The CacheInvalidator is still tested though.